requirements.txt
tavily-python
requests
pydantic-settings
